Adição de propriedades de quantidade e quantidade emprestada na entidade Estoque e atualiza a lógica de empréstimos e disponibilidade no serviço de empréstimo

+++ src/service/EstoqueService.ts
@@ -23,7 +23,7 @@ export class EstoqueService{
             throw new Error("Já existe um exemplar com este código!");
         }
 
-        const livroNoEstoque = new EstoqueEntity(data.isbn, data.cod);
+        const livroNoEstoque = new EstoqueEntity(data.isbn, data.cod, data.quantidade, 0);
         
         this.estoqueRepository.insereLivroNoEstoque(livroNoEstoque);
         
@@ -53,12 +53,23 @@ export class EstoqueService{
     atualizarDisponibilidade(data: any){
         const cod = data.cod;
         const novaDisponibilidade = data.novaDisponibilidade;
+        const estoque = this.estoqueRepository.filtraLivroNoEstoque(cod);
+
+        if(estoque && estoque.quantidade_emprestada === estoque.quantidade){
+            this.estoqueRepository.atualizarDisponibilidade(cod, { disponibilidade: 'não-disponivel'});
+        }
 
         return this.estoqueRepository.atualizarDisponibilidade(cod, novaDisponibilidade);
     }
 
     removerLivroNoEstoque(cod: number){
-        return this.estoqueRepository.removerLivroNoEstoque(cod);
+        const estoque = this.estoqueRepository.filtraLivroNoEstoque(cod);
+        if(estoque && estoque.quantidade === 0){
+            return this.estoqueRepository.removerLivroNoEstoque(cod);
+        }
+        else{
+            throw new Error("Há exemplares emprestados desse Livro, assim que não tiver mais exemplares emprestados você poderá remover este livro no estoque");
+        }
     }
 
 }
